Self is now propagated through all functions as first argument. Accesses to self now refer to the first variable in currentFunction rather than the global variable
All self parameters are passed as ptr to i256 types. This is by design(discussed with kostas).
Reason: These functions can be called by other contracts (external calls) and those contracts might not have access to self type of callee contract.
But callee contract does, so callee casts the self to self_type anytime it encounters access to self.

This can be easily removed later if we decide so